refactor: Nested radio button group in input-choice
