+++ presentation/component/input-choice/input-radio-button-group/_input-radio-button.scss
