+++ presentation/style/game-system.scss
@@ -38,7 +38,7 @@
 @import '../component/label/label';
 @import '../component/lazy-load/lazy-load';
 @import '../component/sortable-list/list-item-order-controls';
-@import '../component/input-radio-button-group/input-radio-button';
+@import '../component/input-choice/input-radio-button-group/input-radio-button';
 @import '../component/input-rich-text/input-rich-text';
 @import '../component/input-number-spinner/input-number-spinner';
 @import '../component/input-textfield/input-textfield';
