+++ presentation/templatePreloader.mjs
@@ -36,9 +36,9 @@ export const TEMPLATES = {
   // Inputs
   COMPONENT_INPUT_TEXTFIELD: `${basePathComponent}/input-textfield/input-textfield.hbs`,
   COMPONENT_INPUT_DROPDOWN: `${basePathComponent}/input-choice/input-dropdown/input-dropdown.hbs`,
+  COMPONENT_INPUT_RADIO_BUTTON_GROUP: `${basePathComponent}/input-choice/input-radio-button-group/input-radio-button-group.hbs`,
   COMPONENT_INPUT_NUMBER_SPINNER: `${basePathComponent}/input-number-spinner/input-number-spinner.hbs`,
   COMPONENT_INPUT_TEXTAREA: `${basePathComponent}/input-textarea/input-textarea.hbs`,
-  COMPONENT_INPUT_RADIO_BUTTON_GROUP: `${basePathComponent}/input-radio-button-group/input-radio-button-group.hbs`,
   COMPONENT_INPUT_IMAGE: `${basePathComponent}/input-image/input-image.hbs`,
   COMPONENT_INPUT_RICH_TEXT: `${basePathComponent}/input-rich-text/input-rich-text.hbs`,
   COMPONENT_INPUT_SEARCH: `${basePathComponent}/input-search/input-search.hbs`,
